Add Card Feature for the result after search
This commit modifies the design of the information retrieved after the
similarity search.
Addition of two files: Search_data.js and Upload_data.js
This files represents the search and upload feature respectively
